WHY ARE ARRAYS SLOW
I tried to optimise the indexing from my rook cache. But no matter what I did, it is always above 4 microseconds which is insanely slow. So now I have given up on making the project fast and decided that its entire purpose of this engine is to say that I have made a chess engine in rust. Also, I reworked the rook magic number system, letting me reduce the number of times I have to index from an array. Also I refactored a lot of the code because the entire purpose of this project is to make the code look nice.